# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Seorilabs Platform API
  version: 1.0.0-draft
  description: |
    Seorilabs 공통 플랫폼 API.

    ## 계약 규칙

    - **이 문서가 계약의 source of truth다.** Go 타입도 TS 타입도 여기서 생성된 산출물이다
    - **`/v1`은 영구히 깨지지 않는다.** 마켓 배포된 구버전 SDK가 2~3년 산다
    - 필드 **추가만** 허용한다. 제거와 의미 변경은 `/v2`
    - 양방향 forward-compat — 서버는 요청의 미지 필드를 무시하고, SDK는 응답의 미지 필드를 무시한다
    - 모든 응답은 envelope로 감싼다. 성공은 `{ok:true, result}`, 실패는 `{ok:false, error}`
    - 요청 본문 상한 64KiB. 제어문자 거부. 허용 필드 외 어떤 키도 400

    상태: **D0 초안.** 스키마 세부는 각 단계에서 확정한다.
servers:
  - url: https://platform-api.seorilabs.example/v1
    description: 확정 필요 — P0 배포 후 실제 URL로 교체
...
paths:
...
  /iap/webhooks/play:
    post:
      tags: [webhook]
      summary: Google Play RTDN 수신
      description: |
        Pub/Sub push subscription. OIDC 토큰으로 인증한다.

        topic은 `play-iap-rtdn` 고정이다. "goog" prefix가 금지되어 이 이름이어야 한다.

        부분환불은 지원하지 않고 422로 거부한다.
        pendingRefundReview는 원장에 기록만 하고 자동 응답하지 않는다.
      operationId: receiveGooglePlayRtdn
      security:
        - PubSubOidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: object
                  required: [data, messageId]
                  properties:
                    data: { type: string, format: byte }
                    messageId: { type: string }
                    publishTime: { type: string, format: date-time }
                subscription: { type: string }
      responses:
        '200':
          description: 처리 또는 무시
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/WebhookOutcome'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    PlatformToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '`/auth/session`이 발급한 플랫폼 세션 토큰'
    PubSubOidc:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Pub/Sub push subscription이 붙이는 Google OIDC 토큰
    AdminOidc:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC ID token
      description: |
        Cloud Run Admin audience를 대상으로 발급한 Google OIDC ID token.
        GET은 read 또는 write allowlist, mutation은 write allowlist만 허용한다.
...
  schemas:
    EnvelopeOk:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    EnvelopeError:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: snake_case 에러 코드. 전체 목록은 docs/03-architecture/iap.md
              example: purchase_owned_by_another_user
            message:
              type: string
              description: 한국어 메시지. 클라이언트가 이 문자열로 분기하지 않는다
...
    WebhookOutcome:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [processed, duplicate, ignored, ignored_invalid, test_notification]
...
  responses:
...
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
...
    Unprocessable:
      description: 구매가 유효하지 않거나 상품 타입이 맞지 않음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
...
    ServiceUnavailable:
      description: 설정 오류이거나 일시적 처리 불가
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
...
security:
  - PlatformToken: []